backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlmodel==0.0.19
pydantic-settings==2.3.1
openai==1.35.0
langchain==0.2.5
logfire>=3.0.0,<5.0.0
python-multipart==0.0.9
httpx==0.27.0